Day 6

+++ src/DayXX.kt
@@ -1,7 +1,7 @@
 import kotlin.time.measureTime
 
 fun main() {
-    val day = object {}::class.java.name.substringBefore("Kt").substringAfter("Day").toInt()
+    val day = object {}::class.java.name.substringBefore("Kt").substringAfter("Day")
 
     fun runParts(input: List<String>, label: String) {
         measureTime { println("$label Part 1: ${part1(input)}") }.also { println("Time Part 1: $it") }
